requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.2
django-annoying==0.10.6
django-model-utils==4.2.0
django-tracking-model==0.1.5
install==1.3.5
mypy-extensions==0.4.3
pathspec==0.10.1
six==1.16.0


